fix compile warning in win

+++ benchmark/thread_trans/benchmark_array_blocking_queue.c
@@ -30,7 +30,7 @@ void run_array_blocking_queue(
 
 	MUGGLE_LOG_INFO("init blocks ok");
 
-	int total_msg_num = num_thread * args->cfg->loop * args->cfg->cnt_per_loop;
+	int total_msg_num = num_thread * (int)args->cfg->loop * (int)args->cfg->cnt_per_loop;
 	muggle_array_blocking_queue_t queue;
 	muggle_atomic_int capacity = total_msg_num / 64;
 	if (muggle_array_blocking_queue_init(&queue, capacity))
